Refactor aoi scene
