+++ src/containers/scenes/aoi-scene/index.js
@@ -1,12 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
-import localforage from 'localforage';
 import orderBy from 'lodash/orderBy';
 import isEmpty from 'lodash/isEmpty';
 import { loadModules } from 'esri-loader';
-import EsriFeatureService from 'services/esri-feature-service';
 import aoisGeometriesActions from 'redux_modules/aois-geometries';
-import { STRINGIFIED_ATTRIBUTES } from 'constants/aois';
+import { setPrecalculatedAOIs, recoverOrCreateNotPrecalculatedAoi } from './aoi-scene-utils';
 
 // actions
 import aoisActions from 'redux_modules/aois';
@@ -14,30 +12,11 @@ import * as urlActions from 'actions/url-actions';
 
 // utils
 import { activateLayersOnLoad, setBasemap } from 'utils/layer-manager-utils';
-import {
-  writeToForageItem
-} from 'utils/local-forage-utils';
-import { calculateGeometryArea } from 'utils/analyze-areas-utils';
-import {
-  getSpeciesData,
-  getContextData,
-  getAoiFromDataBase,
-  getPrecalculatedContextualData,
-  setPrecalculatedSpeciesData,
-} from 'utils/geo-processing-services';
+import { writeToForageItem } from 'utils/local-forage-utils';
 
-// constants
-import {
-  BIRDS,
-  MAMMALS,
-  REPTILES,
-  AMPHIBIANS,
-} from 'constants/geo-processing-services';
 import { layersConfig } from 'constants/mol-layers-configs';
 import { FIREFLY_BASEMAP_LAYER, SATELLITE_BASEMAP_LAYER } from 'constants/layers-slugs';
-import { LAYERS_URLS } from 'constants/layers-urls';
 import { AREA_TYPES } from 'constants/aois';
-import { WDPA_OECM_FEATURE_DATA_LAYER, HALF_EARTH_FUTURE_TILE_LAYER } from 'constants/layers-slugs.js';
 
 // local
 import Component from './component.jsx';
@@ -78,151 +57,17 @@ const AOIScene = props => {
     })
   }, [])
 
-  // PRECALCULATED AOIs
+  // Get PRECALCULATED AOIs
   useEffect(() => {
-    const setPrecalculatedAOIs = () => {
-      EsriFeatureService.getFeatures({
-        url: LAYERS_URLS[precalculatedLayerSlug],
-        whereClause: `MOL_ID = '${aoiId}'`,
-        returnGeometry: true
-      }).then((features) => {
-        const { geometry, attributes } = features[0];
-        setGeometry(geometry);
-
-        const setProtectedAreasType = () => {
-          // Special case for WDPA areas
-          // call to WDPA_OECM_FEATURE_DATA_LAYER with MOL_ID as parameter
-
-          // TODO: Can we get also the geometry from the WDPA_OECM_FEATURE_DATA_LAYER layer?
-          EsriFeatureService.getFeatures({
-            url: LAYERS_URLS[WDPA_OECM_FEATURE_DATA_LAYER],
-            whereClause: `MOL_ID = '${aoiId}'`,
-            returnGeometry: false
-          }).then((results) => {
-            const { attributes: protectedAreaAttributes } = results[0];
-            setContextualData(getPrecalculatedContextualData(protectedAreaAttributes, precalculatedLayerSlug, true, true))
-            setPrecalculatedSpeciesData(protectedAreaAttributes, setTaxaData);
-          });
-        }
-
-        const setNationalOrSubnationalType = () => {
-          setContextualData(getPrecalculatedContextualData(attributes, precalculatedLayerSlug));
-          setPrecalculatedSpeciesData(attributes, setTaxaData);
-        }
-
-        const areaType = setAreaType(attributes);
-        if (areaType === AREA_TYPES.protected) {
-          setProtectedAreasType();
-        } else {
-          setNationalOrSubnationalType();
-        }
-      })
-    }
-
-    // PRECALCULATED FUTURE PLACES
-    const setFuturePlace = () => {
-      changeGlobe({ areaType: AREA_TYPES.futurePlaces })
-
-      EsriFeatureService.getFeatures({
-        url: LAYERS_URLS[HALF_EARTH_FUTURE_TILE_LAYER],
-        whereClause: `OBJECTID = '${objectId}'`,
-        returnGeometry: true
-      }).then((results) => {
-        const { attributes, geometry } = results[0];
-        setPrecalculatedSpeciesData(attributes, setTaxaData);
-        setGeometry(geometry);
-        const areaName = `Priority place ${attributes.cluster}`;
-        setContextualData(getPrecalculatedContextualData({ ...attributes, jsonGeometry: JSON.stringify(geometry), areaName, aoiId }, null, true, true, areaName))
-      });
-    }
-
     if (precalculatedLayerSlug && geometryEngine) {
-      if (areaTypeSelected === AREA_TYPES.futurePlaces || precalculatedLayerSlug === HALF_EARTH_FUTURE_TILE_LAYER) {
-        setFuturePlace()
-      }  else {
-        setPrecalculatedAOIs();
-      }
+      setPrecalculatedAOIs({ areaTypeSelected, precalculatedLayerSlug, aoiId, objectId, setGeometry, setContextualData, setTaxaData, setAreaType, changeGlobe });
     }
   }, [precalculatedLayerSlug, geometryEngine, objectId])
 
-  // NOT PRECALCULATED AOIs
+  // Get NOT PRECALCULATED AOIs
   useEffect(() => {
     if (aoiId && geometryEngine && jsonUtils && !precalculatedLayerSlug) {
-
-      const createNewCustomAOI = () => {
-        const areaName = 'Custom area';
-        const area = calculateGeometryArea(aoiStoredGeometry, geometryEngine);
-        const contextualData = { area, areaName, aoiId, isCustom: true }
-        // Set data before fetch just to show name and available info
-        setContextualData(contextualData);
-
-        const jsonGeometry = aoiStoredGeometry && aoiStoredGeometry.toJSON();
-        setGeometry(jsonUtils.fromJSON(jsonGeometry));
-
-        const forageStoredArea = { ...contextualData, jsonGeometry, timestamp: Date.now() };
-        writeToForageItem(aoiId, forageStoredArea);
-        setStoredArea(forageStoredArea);
-
-        getContextData(aoiStoredGeometry).then((data) => {
-          setContextualData({ ...data, ...contextualData });
-        });
-
-        [BIRDS, MAMMALS, REPTILES, AMPHIBIANS].forEach(taxa => {
-          getSpeciesData(taxa, aoiStoredGeometry).then((data) => {
-            let dataWithoutWhales = data;
-            // WHALES IDS NEED TO BE TEMPORARILY DISCARDED (2954, 2955)
-            if (taxa === MAMMALS) {
-              dataWithoutWhales = dataWithoutWhales.filter((sp) => sp.sliceNumber !== 2954 && sp.sliceNumber !== 2955);
-            }
-            setTaxaData(dataWithoutWhales);
-          });
-        })
-      }
-
-      const recoverAOIFromDB = (aoiData) => {
-        const { geometry, attributes: { species, ...otherAttributes } } = aoiData;
-        setGeometry(geometry);
-        setSpeciesData({ species: JSON.parse(species) });
-        setContextualData({ ...otherAttributes,
-          ...STRINGIFIED_ATTRIBUTES.reduce((acc, key) => {
-            acc[key] = JSON.parse(otherAttributes[key]);
-            return acc;
-          }, {}),
-          percentage: otherAttributes.per_global,
-          isCustom: true
-        })
-      }
-
-      const recoverAOIfromLocal = (localStoredAoi) => {
-        const { jsonGeometry, species, ...rest } = localStoredAoi;
-        const geometry = jsonUtils.fromJSON(jsonGeometry);
-        setGeometry(geometry);
-        setSpeciesData({ species });
-        const contextualData = { ...rest,
-          aoiId, isCustom: true };
-
-        setContextualData(contextualData);
-      }
-
-      localforage.getItem(aoiId).then((localStoredAoi) => {
-        if (localStoredAoi && localStoredAoi.jsonGeometry) {
-          setAreaTypeSelected(AREA_TYPES.custom);
-          recoverAOIfromLocal(localStoredAoi)
-        } else {
-          getAoiFromDataBase(aoiId).then((aoiData) => {
-            if (aoiData && aoiData[0]) {
-              recoverAOIFromDB(aoiData[0])
-            } else {
-              // If we don't have it anywhere we just execute the GP services job to create one
-              createNewCustomAOI()
-            }
-          }).catch(error => {
-            console.error('Could not retrieve AOI', error);
-          })
-        }
-      }).catch((error) => {
-        console.error(error)
-      })
+      recoverOrCreateNotPrecalculatedAoi({ aoiStoredGeometry, geometryEngine, aoiId, jsonUtils, setContextualData, setGeometry, setStoredArea, setTaxaData, setSpeciesData, setAreaTypeSelected });
     }
   }, [aoiId, geometryEngine, jsonUtils])
 
@@ -231,7 +76,7 @@ const AOIScene = props => {
     setSpeciesData({ species: orderedSpecies });
   }, [taxaData])
 
-
+  // Reconcile all data until completely loaded
   useEffect(() => {
     const hasAllData = speciesData && contextualData && !isEmpty(contextualData) && (!contextualData.isCustom || contextualData.protectedAreasList);
     if (!precalculatedLayerSlug && hasAllData) {
@@ -260,6 +105,7 @@ const AOIScene = props => {
   }
 
   const handleFuturePlaceClick = (results) => {
+    if(!results) return;
     const { graphic } = results[0] || {};
     if (!graphic) return;
     const { attributes, geometry } = graphic;
